Refactor: Use enum for Currency, OrderBy, and Direction
