+++ Levi9.ERP.Domain/Models/DTO/DocumentItemDTO.cs
@@ -5,7 +5,7 @@
         public int ProductId { get; set; }
         public string Name { get; set; }
         public float PriceValue { get; set; }
-        public string Currency { get; set; }
+        public CurrencyType Currency { get; set; }
         public int Quantity { get; set; }
     }
 }
