+++ Levi9.ERP.Domain/Models/DocumentType.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Levi9.ERP.Domain.Models
+{
+    public enum DocumentType
+    {
+        INVOICE,
+        RECEIPTS,
+        PURCHASE
+    }
+}
