+++ Levi9.ERP/Datas/Requests/DocumentRequest.cs
@@ -1,4 +1,5 @@
-﻿using Levi9.ERP.Domain.Models.DTO;
+﻿using Levi9.ERP.Domain.Models;
+using Levi9.ERP.Domain.Models.DTO;
 using System.ComponentModel.DataAnnotations;
 
 namespace Levi9.ERP.Datas.Requests
@@ -10,7 +11,7 @@ namespace Levi9.ERP.Datas.Requests
         public int ClientId { get; set; }
         [Required]
         [RegularExpression("^(INVOICE|RECEIPTS|PURCHASE)$", ErrorMessage = "The value of Document type can be: INVOICE, RECEIPTS, or PURCHASE.")]
-        public string DocumentType { get; set; }
+        public DocumentType DocumentType { get; set; }
         [Required]
         public List<DocumentItemDTO> Items { get; set; }
