+++ Levi9.ERP/appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "ErpDatabase": "Data Source=DESKTOP-887L9S3\\SQLEXPRESS;Initial Catalog=erp;Integrated Security=True;TrustServerCertificate=True"
+    "ErpDatabase": "Data Source=localhost;Initial Catalog=Erp; User id = tea; Password = ivona; Integrated Security=True;TrustServerCertificate=True"
   },
   "Logging": {
     "LogLevel": {
